% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_hyperparameters.R
\name{ancestral_reconstruction}
\alias{ancestral_reconstruction}
\title{Ancestral Reconstruction of Spectral Curves}
\usage{

  ancestral_reconstruction(predicted_weights = predictive_distribution()$predictive_mean,
  basis = get_independent_components())
}
\description{
Ancestral Reconstruction of Spectral Curves
}
