% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sampling.R
\name{balanced_samples}
\alias{balanced_samples}
\title{Sampling balanced by species}
\usage{
balanced_samples(df = calls, categorical_variable = calls$species,
  individuals_per_category = 4, n_samples = 10, ignored_variables = c(1,
  2))
}
\description{
Sampling balanced by species
}
