% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bat_hyperparameter_estimates}
\alias{bat_hyperparameter_estimates}
\title{Hyperparameter estimates for Bat Echolocation}
\format{An object of class \code{array} of dimension 4 x 7 x 1000.}
\usage{
bat_hyperparameter_estimates
}
\description{
Hyperparameter estimates for Bat Echolocation
}
\keyword{datasets}
