% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_hyperparameters.R
\name{bat_hyperparameter_estimation}
\alias{bat_hyperparameter_estimation}
\title{Run Phylogenetic Gaussian Process Hyperparameter Estimation}
\usage{
bat_hyperparameter_estimation(df = smooth_and_regularise_call_densities(),
  phylogenetic_tree = phylogeny, basis = get_independent_components(),
  samples_per_mean = 4, n_mean_estimates = 4, n_samples = 10)
}
\description{
Run Phylogenetic Gaussian Process Hyperparameter Estimation
}
