% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sampling.R
\name{calls_per_individual}
\alias{calls_per_individual}
\title{Calls per individual in Bat Echolocation call Dataset}
\usage{
calls_per_individual(df = calls)
}
\arguments{
\item{df}{Defaults to the Bat Echolocation call dataset}
}
\description{
Calls per individual in Bat Echolocation call Dataset
}
