% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCA.R
\name{cubica34}
\alias{cubica34}
\title{CubICA (IMPROVED CUMULANT BASED ICA-ALGORITHM)}
\usage{
[R,y]=cubica34(x)
}
\arguments{
\item{x}{NxP matrix of observations (N: Number of components; P: Number of datapoints(samplepoints))}
}
\value{
R is an NxN matrix such that u=R*x, and u has (approximately) independent components. y is an NxP matrix of independent components
}
\description{
This port of the original cubica34.m written by Pantelis Hadjipantelis.
}
\details{
This algorithm performes ICA by diagonalization of third- and fourth-order cumulants simultaneously.
}
\references{
T. Blaschke and L. Wiskott, "An Improved Cumulant Based Method for Independent Component Analysis", Proc. ICANN-2002, Madrid, Spain, Aug. 27-30.
}
