% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_manipulation.R
\name{expand_tree}
\alias{expand_tree}
\title{Tree expansion}
\usage{
expand_tree(phylogenetic_tree, n_nodes, evolutionary_time = 0)
}
\arguments{
\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}

\item{n_nodes}{The number of nodes to be added to each tip of the tree.}

\item{evolutionary_time}{The distance from the root of the nodes being added to each of the n_nodes tips.}
}
\value{
An object of class 'phylo' with a similar structure to the input \eqn{phylogenetic_tree} but with \eqn{n_nodes} node at each tip.
}
\description{
Expand a phylogenetic tree by adding extra nodes to each existing tip of the tree.
}
