% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPCA.R
\name{get_independent_components}
\alias{get_independent_components}
\title{Get Independent Components}
\usage{
get_independent_components(principal_components = spectral_fpca)
}
\description{
Get Independent Components
}
