% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sampling.R
\name{get_mean_estimates}
\alias{get_mean_estimates}
\title{Sample of Species Mean Estimates}
\usage{
get_mean_estimates(df = smooth_and_regularise_call_densities(),
  samples_per_mean = 4, n_mean_estimates = 4)
}
\description{
Sample of Species Mean Estimates
}
