% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_hyperparameters.R
\name{linear_coefficients}
\alias{linear_coefficients}
\title{Return coefficients for linear model}
\usage{
linear_coefficients(response, predictors)
}
\description{
Return coefficients for linear model
}
