% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_spectralCurve.R
\name{log_spectral_density}
\alias{log_spectral_density}
\title{Log Spectral Density Estimation}
\usage{
log_spectral_density(signal)
}
\arguments{
\item{signal}{A univariate time series}
}
\value{
log10 of the spectral density
}
\description{
Estimates log10 of the spectral density of a signal.
}
