% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_spectralCurve.R
\name{log_spectral_density_calls}
\alias{log_spectral_density_calls}
\title{Log Spectral Density Estimation}
\usage{
log_spectral_density_calls(signals = calls)
}
\arguments{
\item{signals}{Defaults to the echolocation call dataset provided}
}
\value{
log10 of the spectral density for each call.
}
\description{
Applies the 'log_spectral_density()' function to the echolocation call
dataset
}
\seealso{
\code{\link{log_spectral_density}}
}
