% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_ou_simulation.R
\name{phylo_ou_cov}
\alias{phylo_ou_cov}
\title{Phylogenetic Ornstein-Uhlenbeck Process Covariance}
\usage{
phylo_ou_cov(hyperparameters, phylogenetic_tree)
}
\arguments{
\item{hyperparameters}{Vector of hyperparameters for an Ornstein-Uhlenbeck
kernel in form \eqn{(\sigma_p, \ell, \sigma_n)}}

\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}
}
\value{
Outputs the covariance between obsevations taken at the tips of the
  phylogenetic tree.
}
\description{
Derives the Covariance matrix for points in a univariate Phylogenetic
Gaussian Process given the Ornsetin-Uhelenbeck process kernel. \eqn{k(x, x')
= \sigma_p^2 \exp(\frac{|x - x'|}{\ell})\) + \sigma_n^2}
}
\examples{
 K <- phylo_ou_covar(c(1,20,1), phylogeny)
}
