% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_ou_simulation.R
\name{phylo_ou_simulated_obs}
\alias{phylo_ou_simulated_obs}
\title{Simulated Realisations of a Phylogenetic Ornstein-Uhlenbeck Process}
\usage{
phylo_ou_simulated_obs(hyperparameters, phylogenetic_tree, nsamples)
}
\arguments{
\item{hyperparameters}{Vector of hyperparameters for an Ornstein-Uhlenbeck
kernel in form \eqn{(\sigma_p, \ell, \sigma_n)}}

\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}

\item{nsamples}{The number of samples to be simulated.}
}
\value{
Outputs a matrix of size \eqn{ntips x nsamples} containing
  \eqn{nsamples} somulations from a given Phylogenetic Ornstein-Uhlenbeck
  Gaussian Process.
}
\description{
Title says it all really.
}
