% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phylogeny}
\alias{phylogeny}
\title{Phylogenetic Tree for selected Bat species}
\format{List of type 'phylo'. Requires package 'ape'.}
\source{
A. Collen. The evolution of echolocation in bats: a comparative
  approach. PhD thesis, UCL (University College London) (2012).
}
\usage{
phylogeny
}
\description{
Tree encoding phylogenetic relationships between the species of bat recorded
in the 'calls' dataset. Branch lengths set such that 1 unit = 1 million
years.
}
\keyword{datasets}
