% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_ou_simulation.R
\name{pou_covariance}
\alias{pou_covariance}
\title{Phylogenetic Covariance}
\usage{
pou_covariance(ln_hyperparameters, phylogenetic_tree)
}
\arguments{
\item{ln_hyperparameters}{Vector of the natural logarithm of hyperparameters
for an Ornstein-Uhlenbeck kernel in order \eqn{log(phylogenetic noise,
phylogenetic length-scale, non-phylogenetic noise)}}

\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}
}
\value{
Outputs the covariance between obsevations at the tips of the
  phylogenetic tree.
}
\description{
Derives the Covariance matrix for points in a univariate Phylogenetic
Gaussian Process given the Ornsetin-Uhelenbeck process kernel.
}
\examples{
 K <- pou_cov(log(c(1,20,1)), phylogeny)

}
