% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{pou_logl_fast}
\alias{pou_logl_fast}
\title{Fast Phylogenetic Likelihood}
\source{
P. Z. Hadjipantelis, N. S. Jones, J. Moriarty, D. A. Springate, and
  C. G. Knight, Function-valued traits in evolution, Journal of The Royal
  Society Interface. 10(82), 20121032 (2013). \url{https://github.com/fpgpr}
}
\usage{
pou_logl_fast(ln_hyperparameters, phylogenetic_tree, observations)
}
\arguments{
\item{ln_hyperparameters}{Vector of the natural logarithm of hyperparameters
for an Ornstein-Uhlenbeck kernel in order \eqn{log(phylogenetic noise,
phylogenetic length-scale, non-phylogenetic noise)}}

\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}

\item{observations}{Observations from the tips of the phylogenetic tree}
}
\value{
A positive value which is the negative log likelihood for the data.
}
\description{
Using the inline and RcppEigen packages performs a fast calculation of the
Phylogenetic Ornsetin-Uhelenbeck Process negative log likelihood, given the
process hyperparameters, a phylogenetic tree, and some data. The C++ code was
taken from code accompanying Hajipantelis et al.
}
