% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{pou_logl_slow}
\alias{pou_logl_slow}
\title{Phylogenetic Likelihood}
\usage{
pou_logl_slow(ln_hyperparameters, phylogenetic_tree, observations)
}
\arguments{
\item{ln_hyperparameters}{Vector of the natural logarithm of hyperparameters
for an Ornstein-Uhlenbeck kernel in order \eqn{log(phylogenetic noise,
phylogenetic length-scale, non-phylogenetic noise)}}

\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}

\item{observations}{Observations from the tips of the phylogenetic tree}
}
\value{
A positive value which is the negative log likelihood for the data.
}
\description{
Calculates the Phylogenetic Ornstein-Uhelenbeck Process negative log
likelihood, given the process hyperparameters, a phylogenetic tree, and some
data.
}
