% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_ou_simulation.R
\name{pou_simulated_data}
\alias{pou_simulated_data}
\title{Simulate Phylogenetic Data}
\usage{
pou_simulated_data(ln_hyperparameters, phylogenetic_tree, n_samples)
}
\arguments{
\item{nsamples}{The number of samples to be simulated.}
}
\value{
Outputs a matrix of size \eqn{ntips x nsamples} containing
  \eqn{nsamples} simulations from a given Phylogenetic Ornstein-Uhlenbeck
  Gaussian Process.
}
\description{
Simulated Realisations of a Phylogenetic Ornstein-Uhlenbeck Process
}
