% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2_mle.R
\name{pou_type2mle}
\alias{pou_type2mle}
\title{Hyperparameter Estimation}
\usage{
pou_type2mle(phylogenetic_tree, observations, logl_function = pou_logl_slow,
  optim_function = "optim", optim_method = "Nelder-Mead",
  lower_initialisation = c(0, 0, 0), upper_initialisation = c(1, 1, 1),
  n_restarts = 1)
}
\arguments{
\item{phylogenetic_tree}{Phylogenetic tree of relationships between
observations}

\item{observations}{Observations from the tips of the phylogenetic tree}

\item{logl_function}{function calculating the negative log likelihood of the
observations to be minimised. Either 'pou_logl_slow' or 'pou_logl_fast'.}

\item{optim_function}{String indicating which optimising function to use,
'optim' or 'uobyqa' from the minqa package.}

\item{optim_method}{Method to be used with the optim optimiser.}

\item{lower_initialisation}{Lower bound for the uniform random variable used
to generate initial values.}

\item{upper_initialisation}{Upper bound for the uniform random variable used
to generate initial values.}

\item{n_restarts}{Number of different initialisation values to test.}
}
\value{
Vector of optimised hyperparameters and the corresponding negative
  log likelihood, in order (phylogenetic noise, phylogenetic length-scale,
  non-phylogenetic noise, negative log likelihood)
}
\description{
Type II Maximum Likelihood Estimation for hyperparameters of a Phylogenetic
Ornstein-Uhlenbeck Process.
}
