% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_hyperparameters.R
\name{representative_basis_weights}
\alias{representative_basis_weights}
\title{Representative weights for each species}
\usage{
representative_basis_weights(representative_curves = species_mean(),
  basis = get_independent_components())
}
\description{
Representative weights for each species
}
