% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_spectralCurve.R
\name{rescale_density}
\alias{rescale_density}
\title{Scale curvea}
\usage{
rescale_density(spectral_density)
}
\arguments{
\item{spectral_density}{curve to be mapped to interval}
}
\value{
Curve on interval [0, 1]
}
\description{
Rescales a curve such that it lies along the interval [0, 1].
}
