% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_spectralCurve.R
\name{rescale_smooth_densities}
\alias{rescale_smooth_densities}
\title{Curve Scaling}
\usage{

  rescale_smooth_densities(spectral_densities = smooth_and_regularise_call_densities())
}
\arguments{
\item{spectral_densities}{Defaults to the smoothed and reguarised spectral
densities of the echolocation call dataset provided.}
}
\value{
smoothed and regularised spectral density for each call scaled to the
  interval [0, 1].
}
\description{
Applies the 'rescale_dinsity' function to the smoothed and regularised
spectral densities of the echolocation call dataset.
}
\seealso{
\code{\link{log_spectral_density}}
  \code{\link{log_spectral_density_calls}}
  \code{\link{smooth_and_regularise}}
  \code{\link{smooth_and_regularise_densities}} \code{\link{rescale_density}}
}
