% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sampling.R
\name{sample_mean_estimates}
\alias{sample_mean_estimates}
\title{Resampled estimates of Species mean}
\usage{
sample_mean_estimates(df = smooth_and_regularise_call_densities(),
  samples_per_mean = 4, n_mean_estimates = 4, n_samples = 10)
}
\description{
Resampled estimates of Species mean
}
