% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_to_spectralCurve.R
\name{smooth_and_regularise}
\alias{smooth_and_regularise}
\title{Curve Smoothing and Regularisation}
\usage{
smooth_and_regularise(spectrum, observed_range = c(0, 250),
  regularised_range = c(9, 212), n_points = 208)
}
\arguments{
\item{observed_range}{The range over which the noisy input curve is observed at equidistant points.}

\item{regularised_range}{The range over which the smoothed representation of the curve is to be regularised to.}

\item{n_points}{Number of equidistant points along the x axis to evaluate the
smoothed curve at.}
}
\value{
A smoothed and regularised representation of the input.
}
\description{
Fits smoothing splines to noisy observations of a curve and then regularises
the curve by interpolating curve values over \eqn{n_points} equidistant
points on along the x-axis.
}
