% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative_curves.R
\name{species_mean}
\alias{species_mean}
\title{Representative spectral curves}
\usage{
species_mean(df = smooth_and_regularise_call_densities(),
  ignored_columns = c(1, 2), vector_length = length(df$chirp1[[1]]))
}
\description{
Representative spectral curves
}
