% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectral_fpca}
\alias{spectral_fpca}
\title{Spectral Components}
\format{An object of class \code{matrix} with 208 rows and 6 columns.}
\usage{
spectral_fpca
}
\description{
Approximate Functional Principal Components of the Bat Echolocation Call
Energy Spectral Densities obtained by a resampling of call datasets balanced
by both species and individual and grouping components by peak frequency.
}
\keyword{datasets}
