% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spectral_mean}
\alias{spectral_mean}
\title{Spectral Mean of Bat Echolocation Calls}
\format{An object of class \code{numeric} of length 208.}
\usage{
spectral_mean
}
\description{
Approximate mean Energy Spectral density for Bat Echolocation calls generated
by a resampling of call datasets balanced by both species and individual.
}
\keyword{datasets}
