% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{pou_logl_slow}
\alias{pou_logl_slow}
\title{Phylogenetic Likelihood}
\usage{
pou_logl_slow(ln_hyperparameters, phylogenetic_tree, observations)
}
\arguments{
\item{dat}{Observations from the tips of the phylogenetic tree}
}
\value{
A positive value which is the negative log likelihood for the data.
}
\description{
Calculates the Phylogenetic Ornsetin-Uhelenbeck Process negative log
likelihood, given the process hyperparameters, a phylogenetic tree, and some
data.
}
